rpi/requirements.txt
pyserial==3.5
yfinance==0.2.40
